+++ src/plugins/Display/Display_ePaper.cpp
@@ -26,8 +26,8 @@ DisplayEPaper::DisplayEPaper() {
 void DisplayEPaper::init(uint8_t type, uint8_t _CS, uint8_t _DC, uint8_t _RST, uint8_t _BUSY, uint8_t _SCK, uint8_t _MOSI, uint32_t *utcTs, const char *version) {
     mUtcTs = utcTs;
 
-    mRefreshState = RefreshStatus::LOGO;
-    mSecondCnt = 0;
+    mRefreshState = RefreshStatus::BLACK;
+    mSecondCnt = 2;
 
     if (type == 10) {
         Serial.begin(115200);
@@ -63,6 +63,7 @@ void DisplayEPaper::fullRefresh() {
 void DisplayEPaper::refreshLoop() {
     switch(mRefreshState) {
         case RefreshStatus::LOGO:
+            mFirst = false;
             _display->fillScreen(GxEPD_BLACK);
             _display->drawBitmap(0, 0, logo, 200, 200, GxEPD_WHITE);
             mNextRefreshState = RefreshStatus::PARTITIALS;
@@ -79,7 +80,7 @@ void DisplayEPaper::refreshLoop() {
             if(mSecondCnt == 0) {
                 _display->fillScreen(GxEPD_WHITE);
                 mNextRefreshState = RefreshStatus::PARTITIALS;
-                mRefreshState = RefreshStatus::WAIT;
+                mRefreshState = (mFirst) ? RefreshStatus::LOGO : RefreshStatus::WAIT;
             }
             break;