+++ src/plugins/Display/Display_ePaper.h
@@ -58,6 +58,7 @@ class DisplayEPaper {
         const char* _version;
         RefreshStatus mRefreshState, mNextRefreshState;
         uint8_t mSecondCnt;
+        bool mFirst = true;
 };
 
 #endif  // ESP32